% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{readSIXEL}
\alias{readSIXEL}
\title{Read a SIXEL image}
\usage{
readSIXEL(source)
}
\arguments{
\item{source}{character, name of the file to read from.}
}
\value{
A raster array with values ranging from 0 to 1.
The array has dimensions (height, width, 3) where the third dimension
represents the R, G, and B color channels.
}
\description{
Reads an image from a SIXEL file into a raster array.
}
\examples{
# read a sample file
img <- readSIXEL(system.file("snake.six", package="rsixel"))

}
