% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode.R
\name{sixelDecode}
\alias{sixelDecode}
\title{Decode SIXEL escape sequence to image data}
\usage{
sixelDecode(data)
}
\arguments{
\item{data}{character, SIXEL escape sequence.}
}
\value{
A three dimensional RGB array with values ranging from 0 to 1.
  The array has dimensions [height, width, 3] where the third dimension
  represents the R, G, and B color channels.
}
\description{
Parse a SIXEL escape sequence and convert it to a three dimensional RGB array.
}
\examples{
# Encode an image to SIXEL and decode it back
img <- jpeg::readJPEG(system.file("img", "Rlogo.jpg", package="jpeg"))
sixel_data <- sixelEncode(img, 4)
decoded <- sixelDecode(sixel_data)

}
